requirements.txt
Django>=4.2
psycopg2-binary
docling
datasets
python-dotenv